Isolate generic OpenAPI MCP tool handlers
